Conflicts:
	dspace/config/log4j.xml

+++ dspace/modules/dataone-mn/dataone-mn-webapp/src/main/java/org/dspace/dataonemn/ListObjects.java
@@ -17,10 +17,10 @@ public class ListObjects extends Element implements Constants{
 		addAttribute(new Attribute("start", Integer.toString(aStart)));
 	}
 	
-	public void setTotal(int aTotal) {
-		addAttribute(new Attribute("total", Integer.toString(aTotal)));
+	public void setTotal(long aTotal) {
+		addAttribute(new Attribute("total", Long.toString(aTotal)));
 	}
-	
+        
 	public String toString() {
 		return toXML();
 	}
